replace old bokeh renderer with updated renderer
