+++ pixiedust/display/chart/renderers/bokeh/bkHistogramDisplay.py
